+++ indico/modules/events/registration/notifications.py
@@ -7,7 +7,7 @@
 
 from __future__ import unicode_literals
 
-from flask import session
+from flask import request, session
 
 from indico.core import signals
 from indico.core.notifications import make_email, send_email
@@ -39,7 +39,8 @@ def _notify_registration(registration, template, to_managers=False):
             registration.state == RegistrationState.complete):
         attachments = get_ticket_attachments(registration)
 
-    template = get_template_module('events/registration/emails/{}'.format(template), registration=registration)
+    template = get_template_module('events/registration/emails/{}'.format(template),
+                                   registration=registration, notify_reason=request.form.get('notify_reason') == 'y')
     to_list = registration.email if not to_managers else registration.registration_form.manager_notification_recipients
     from_address = registration.registration_form.sender_address if not to_managers else None
     mail = make_email(to_list=to_list, template=template, html=True, from_address=from_address, attachments=attachments)